backend/pyproject.toml
[tool.poetry]
name = "citadels-api"
version = "0.1.0"
description = ""
authors = ["vinalina <vinalian23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
loguru = "^0.7.2"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
